carts: audioManager: serviceDurationRequest: use mediaStatusChanged
Rather than listening for the durationChanged signal, use the
mediaStatusChanged signal and await for the LoadedMedia status as the
duration could potentially be the same between two carts.  If that was
the case we wouldn't recieve a signal to call the cart callback and we
wouldn't make any progress loading the cart wall.